Move grpc server code to new file
