+++ grpc.go
@@ -56,39 +56,6 @@ func (gp *GRPCFetchProtocol) NewFetcher(address string) (RemoteFetcher, error) {
 	return &grpcFetcher{address: address, conn: conn, client: client}, nil
 }
 
-// GRPCGalaxyCacheServer implements the GalaxyCacheServer
-// interface generated by the GalaxyCache pb service
-type GRPCGalaxyCacheServer struct {
-	universe *Universe
-}
-
-// RegisterGRPCServer registers the given grpc.Server with
-// a Universe for GetFromPeer calls over RPC
-func RegisterGRPCServer(universe *Universe, grpcServer *grpc.Server) {
-	pb.RegisterGalaxyCacheServer(grpcServer, &GRPCGalaxyCacheServer{universe: universe})
-}
-
-// GetFromPeer implements the generated GalaxyCacheServer
-// interface, making an internal Get() after receiving a
-// remote call from a peer
-func (gp *GRPCGalaxyCacheServer) GetFromPeer(ctx context.Context, req *pb.GetRequest) (*pb.GetResponse, error) {
-	group := gp.universe.GetGalaxy(req.Galaxy)
-	if group == nil {
-		// log.Warnf("Unable to find group [%s]", req.Group)
-		return nil, fmt.Errorf("Unable to find group [%s]", req.Galaxy)
-	}
-
-	group.Stats.ServerRequests.Add(1) // keep track of the num of req
-
-	var value ByteCodec
-	err := group.Get(ctx, req.Key, &value)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to retrieve [%s]: %v", req, err)
-	}
-
-	return &pb.GetResponse{Value: value}, nil
-}
-
 // Fetch here implements the RemoteFetcher interface for
 // sending Gets to peers over an RPC connection
 func (g *grpcFetcher) Fetch(ctx context.Context, galaxy string, key string) ([]byte, error) {
